multerModuleFinished

+++ controllers/admin.js
@@ -1,7 +1,7 @@
 
 const Product  = require('../models/product');
 const {validationResult} = require("express-validator")
-
+const fileHelper = require("../util/file")
 exports.getAddProduct = (req, res, next) => {
 
     res.render("admin/edit-product", {
@@ -125,6 +125,7 @@ exports.postEditProduct = (req,res,next) => {
         product.price = updatedPrice;
         product.description = updatedDescription;
         if(image){
+            fileHelper.deleteFile(product.imageUrl);
             product.imageUrl = image.path;
         }
         return product.save()
@@ -141,8 +142,13 @@ exports.postEditProduct = (req,res,next) => {
 };
 exports.postDeleteProduct = (req,res,next) => {
     const prodId = req.body.productId;
-    Product.deleteOne({_id:prodId, userId : req.user._id})
-    .then( () => {
+    Product.findById(prodId).then(product=>{
+        if(!product){
+            return next(new Error("Product not found"))
+        }
+       fileHelper.deleteFile(product.imageUrl);
+       return Product.deleteOne({_id:prodId, userId : req.user._id})
+    }).then( () => {
         console.log("DESTROYED PRODUCT")
         res.redirect('/admin/products');
     })
@@ -150,7 +156,7 @@ exports.postDeleteProduct = (req,res,next) => {
         const error = new Error(err);
         error.httpStatusCode = 500;
         return next(error);
-    })
+    })  
 };
 exports.getProductsAdmin = (req,res,next) => {
     Product.find({userId: req.user._id})
